requirements.txt
allure-pytest==2.13.4
allure-python-commons==2.13.4
assertpy==1.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
curlify==2.2.1
exceptiongroup==1.2.0
Faker==24.4.0
future-fstrings==1.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
networkx==3.2.1
outcome==1.3.0.post0
packaging==23.2
pillow==10.3.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
pytest==8.0.2
pytest-depends==1.0.1
pytest-html==4.1.1
pytest-html-reporter==0.2.9
pytest-metadata==3.1.1
pytest-report==0.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0